+++ api/impl/cactusDiskPrivate.h
@@ -85,11 +85,11 @@ Name cactusDisk_addString(CactusDisk *cactusDisk, const char *string);
  * Retrieves a string from the bucket of sequence.
  */
 char *cactusDisk_getString(CactusDisk *cactusDisk, Name name,
-        int32_t start, int32_t length, int32_t strand, int32_t totalSequenceLength);
+        int64_t start, int64_t length, int64_t strand, int64_t totalSequenceLength);
 
 /*
  * Gets the string from a cache.
  */
-char *cactusDisk_getStringFromCache(CactusDisk *cactusDisk, Name name, int32_t start, int32_t length, int32_t strand);
+char *cactusDisk_getStringFromCache(CactusDisk *cactusDisk, Name name, int64_t start, int64_t length, int64_t strand);
 
 #endif
