+++ api/impl/cactusFacePrivate.h
@@ -14,7 +14,7 @@ struct _face {
 	Flower * flower;
 
 	// Number of top nodes in face, length of subsequent arrays
-	int32_t cardinal;
+	int64_t cardinal;
 
 	// Array of pointers of top nodes of the face
 	Cap **topNodes;
@@ -25,7 +25,7 @@ struct _face {
 
 	// Number of lifted edges which are colinear to the
 	// corresponding top node's adjacency edge
-	int32_t *bottomNodeNumbers;
+	int64_t *bottomNodeNumbers;
 
 	// Destination of the isolated lifted edge
 	Cap ***derivedEdgeDestinations;
@@ -36,7 +36,7 @@ struct _face {
 
 struct _face_FaceEndIterator {
 	Face *face;
-	int32_t index;
+	int64_t index;
 };
 
 ////////////////////////////////////////////////
@@ -55,52 +55,52 @@ Face * face_construct(Flower * flower);
 /*
  * Get selected derived destinations for selected top node
  */
-Cap * face_getDerivedDestinationAtIndex(Face * face, int32_t topIndex, int32_t derivedIndex);
+Cap * face_getDerivedDestinationAtIndex(Face * face, int64_t topIndex, int64_t derivedIndex);
 
 /*
  * Get non-null derived destination of selected top node (useful for simple faces)
  */
-Cap * face_getDerivedDestination(Face * face, int32_t index);
+Cap * face_getDerivedDestination(Face * face, int64_t index);
 
 /*
  * Get the number of bottom nodes for the selected top node
  */
-int32_t face_getBottomNodeNumber(Face * face, int32_t topIndex);
+int64_t face_getBottomNodeNumber(Face * face, int64_t topIndex);
 
 /*
  * Get selected bottom node from selected top node in face
  */
-Cap * face_getBottomNode(Face * face, int32_t topNodeIndex, int32_t bottomNodeIndex);
+Cap * face_getBottomNode(Face * face, int64_t topNodeIndex, int64_t bottomNodeIndex);
 
 /*
  * Allocate arrays to allow for data
  */
-void face_allocateSpace(Face * face, int32_t cardinal);
+void face_allocateSpace(Face * face, int64_t cardinal);
 
 /*
  * Sets the selected top node
  */
-void face_setTopNode(Face * face, int32_t topIndex, Cap * topNode);
+void face_setTopNode(Face * face, int64_t topIndex, Cap * topNode);
 
 /*
  * Set bottom node count and allocate space to store pointers
  */
-void face_setBottomNodeNumber(Face * face, int32_t topIndex, int32_t number);
+void face_setBottomNodeNumber(Face * face, int64_t topIndex, int64_t number);
 
 /*
  * Sets the derived edge destination for a given top node in face
  */
-void face_setDerivedDestination(Face * face, int32_t topIndex, int32_t bottomIndex, Cap * destination);
+void face_setDerivedDestination(Face * face, int64_t topIndex, int64_t bottomIndex, Cap * destination);
 
 /*
  * Adds bottom node to selected top node in face
  */
-void face_addBottomNode(Face * face, int32_t topIndex, Cap * bottomNode);
+void face_addBottomNode(Face * face, int64_t topIndex, Cap * bottomNode);
 
 /*
  * Adds a top node with a single bottom node
  */
-void face_engineerArtificialNodes(Face * face, Cap * topNode, Cap * bottomNode, int32_t nonDerived);
+void face_engineerArtificialNodes(Face * face, Cap * topNode, Cap * bottomNode, int64_t nonDerived);
 
 /*
  * Gets the face end associated with the top node of the cap. The public
